backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
openpyxl==3.1.5
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
langchain-google-genai==2.0.4
langchain-groq==0.2.0
langchain-cerebras==0.4.0
supabase==2.7.4
python-docx==1.1.2
reportlab==4.2.2
